+++ src/app/shared/models/aql/aql-parameter.interface.ts
@@ -28,4 +28,5 @@ export interface IAqlParameter {
   options?: IDictionary<any, any>
   valueType?: AqlParameterValueType
   isMetaFetched?: boolean
+  isDisabled: boolean
 }